Refactored hooks to own directory
Lint fix
